+++ src/tests/commonTests.ts
@@ -21,25 +21,18 @@ function getTag(tagOrHTML: string): CalciteComponentTag {
   if (isHTML(tagOrHTML)) {
     const regex = /[>\s]/;
     const trimmedTag = tagOrHTML.trim();
-    return trimmedTag.substring(
-      1,
-      trimmedTag.search(regex)
-    ) as CalciteComponentTag;
+    return trimmedTag.substring(1, trimmedTag.search(regex)) as CalciteComponentTag;
   }
 
   return tagOrHTML as CalciteComponentTag;
 }
 
-async function simplePageSetup(
-  componentTagOrHTML: TagOrHTML
-): Promise<E2EPage> {
+async function simplePageSetup(componentTagOrHTML: TagOrHTML): Promise<E2EPage> {
   const componentTag = getTag(componentTagOrHTML);
 
   return newE2EPage({
-    html: isHTML(componentTagOrHTML)
-      ? componentTagOrHTML
-      : `<${componentTag}><${componentTag}/>`,
-    failOnConsoleError: true,
+    html: isHTML(componentTagOrHTML) ? componentTagOrHTML : `<${componentTag}><${componentTag}/>`,
+    failOnConsoleError: true
   });
 }
 
@@ -56,10 +49,7 @@ export async function accessible(componentTagOrHTML: TagOrHTML): Promise<void> {
   ).toHaveNoViolations();
 }
 
-export async function renders(
-  componentTagOrHTML: TagOrHTML,
-  invisible?: true
-): Promise<void> {
+export async function renders(componentTagOrHTML: TagOrHTML, invisible?: true): Promise<void> {
   const page = await simplePageSetup(componentTagOrHTML);
   const element = await page.find(getTag(componentTagOrHTML));
 
@@ -80,7 +70,7 @@ export async function reflects(
 
   for (const propAndValue of propsToTest) {
     const { propertyName, value } = propAndValue;
-    const componentAttributeSelector = `${componentTag}[${propertyName}]`;
+    const componentAttributeSelector = `${componentTag}[${propToAttr(propertyName)}]`;
 
     element.setProperty(propertyName, value);
     await page.waitForChanges();
@@ -99,6 +89,10 @@ export async function reflects(
       expect(await page.find(componentAttributeSelector)).toBeTruthy();
     }
   }
+}
+
+function propToAttr(name: string): string {
+  return name.replace(/([a-z0-9])([A-Z])/g, "$1-$2").toLowerCase();
 }
 
 export async function defaults(
@@ -126,4 +120,14 @@ export async function hidden(componentTagOrHTML: TagOrHTML): Promise<void> {
   await page.waitForChanges();
 
   expect(await element.isVisible()).toBe(false);
+}
+
+export async function focusable(componentTagOrHTML: TagOrHTML): Promise<void> {
+  const page = await simplePageSetup(componentTagOrHTML);
+  const tag = getTag(componentTagOrHTML);
+  const element = await page.find(tag);
+
+  await element.callMethod("setFocus"); // assumes element is CalciteFocusableElement
+
+  expect(await page.evaluate(() => document.activeElement.tagName)).toEqual(tag.toUpperCase());
 }