+++ stencil.config.ts
@@ -10,21 +10,13 @@ export const config: Config = {
     { components: ["calcite-button"] },
     { components: ["calcite-card"] },
     { components: ["calcite-chip"] },
+    { components: ["calcite-color", "calcite-color-hex-input", "calcite-color-swatch"] },
     { components: ["calcite-combobox"] },
     {
-      components: [
-        "calcite-date",
-        "calcite-date-month",
-        "calcite-date-month-header",
-        "calcite-date-day",
-      ],
+      components: ["calcite-date", "calcite-date-month", "calcite-date-month-header", "calcite-date-day"]
     },
     {
-      components: [
-        "calcite-dropdown",
-        "calcite-dropdown-group",
-        "calcite-dropdown-item",
-      ],
+      components: ["calcite-dropdown", "calcite-dropdown-group", "calcite-dropdown-item"]
     },
     { components: ["calcite-icon"] },
     { components: ["calcite-input"] },
@@ -43,15 +35,10 @@ export const config: Config = {
     { components: ["calcite-stepper", "calcite-stepper-item"] },
     { components: ["calcite-switch"] },
     {
-      components: [
-        "calcite-tab",
-        "calcite-tab-title",
-        "calcite-tab-nav",
-        "calcite-tabs",
-      ],
+      components: ["calcite-tab", "calcite-tab-title", "calcite-tab-nav", "calcite-tabs"]
     },
     { components: ["calcite-tooltip", "calcite-tooltip-manager"] },
-    { components: ["calcite-tree", "calcite-tree-item"] },
+    { components: ["calcite-tree", "calcite-tree-item"] }
   ],
   outputTargets: [
     { type: "dist-hydrate-script" },
@@ -64,28 +51,28 @@ export const config: Config = {
       prerenderConfig: "./prerender.config.js",
       copy: [{ src: "demos", dest: "demos" }],
       serviceWorker: {
-        unregister: true,
-      },
-    },
+        unregister: true
+      }
+    }
   ],
   globalStyle: "src/assets/styles/global.scss",
   plugins: [
     sass({
-      injectGlobalPaths: ["src/assets/styles/includes.scss"],
-    }),
+      injectGlobalPaths: ["src/assets/styles/includes.scss"]
+    })
   ],
   testing: {
     moduleNameMapper: {
-      "^/assets/(.*)$": "<rootDir>/src/tests/iconPathDataStub.ts",
+      "^/assets/(.*)$": "<rootDir>/src/tests/iconPathDataStub.ts"
     },
-    setupFilesAfterEnv: ["<rootDir>/src/tests/setupTests.ts"],
+    setupFilesAfterEnv: ["<rootDir>/src/tests/setupTests.ts"]
   },
   hydratedFlag: {
     selector: "attribute",
-    name: "calcite-hydrated",
+    name: "calcite-hydrated"
   },
   extras: {
     appendChildSlotFix: true,
-    slotChildNodesFix: true,
-  },
+    slotChildNodesFix: true
+  }
 };